modify ff_port_cfg.port_id's type from uint8_t to unint16_t.

+++ lib/ff_config.h
@@ -61,7 +61,7 @@ struct ff_hw_features {
 struct ff_port_cfg {
     char *name;
     char *ifname;
-    uint8_t port_id;
+    uint16_t port_id;
     uint8_t mac[6];
     struct ff_hw_features hw_features;
     char *addr;
